# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: SPIMS Mobile API
  version: '1.0'
  description: 'Bearer-token API for the SPIMS student and instructor mobile clients. Every registered
    `api.v1.*` route must have an entry here — enforced by tests/Feature/Api/V1/OpenApiCoverageTest.php.

    '
servers:
- url: /api/v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
paths:
...
  /offerings/{offering}/weeks/{week}/complete:
    post:
      operationId: completeWeek
      summary: Mark a week complete.
      security:
      - bearerAuth: []
      parameters:
      - in: path
        name: offering
        required: true
        schema:
          type: string
      - in: path
        name: week
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
        '403':
          description: Not enrolled
...